+++ app/src/main/res/values-night/strings.xml
@@ -18,4 +18,5 @@
     <string name="proventos">Proventos</string>
     <string name="inss">INSS</string>
     <string name="delete">Deletar</string>
+    <string name="save">Salvar</string>
 </resources>
