+++ app/src/main/res/values/strings.xml
@@ -21,4 +21,5 @@
     <string name="proventos">Proventos</string>
     <string name="inss">INSS</string>
     <string name="delete">Deletar</string>
+    <string name="save">Salvar</string>
 </resources>
